syncing, sprites, server fixes, functionality, finished SP

+++ Core/Items/PoTGlobalItem.IO.cs
@@ -76,7 +76,7 @@ partial class PoTGlobalItem : GlobalItem
 		writer.Write(data.SpecialName);
 		writer.Write((byte)data.RealLevel);
 
-		writer.Write(data.Affixes.Count);
+		writer.Write((byte)data.Affixes.Count);
 		foreach (ItemAffix affix in data.Affixes)
 		{
 			affix.NetSend(writer);
